# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Binar Car Management API",
    "contact": {
      "email": "safiraadty@gmail.com"
    }
  },
  "servers": [{
    "url": "http://localhost:3000",
    "description": "Car Management API"
  }],
...
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Show All Users Data",
        "description": "This endpoint for show all users data",
        "operationId": "getUsers",
        "consumes": [
          "application/json"
        ],
        "parameters": [{
          "$ref": "#/components/parameters/id"
        }],
        "responses": {
          "201": {
            "description": "Show All Users Data",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Succesfully show all users data"
                }
              }
            }
          },
          "404": {
            "description": "Route not found!",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Failed"
                },
                "message": {
                  "type": "string",
                  "example": "Route not found!"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "carAuth": {
        "description": "Authentication for Car Management API",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Car id for spesific car",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 100
        }
      }
    }
  },
...
}